template Code: refactor
